Minim : Fixed glitchy lockscreen when entering lockscreen very fast
Repro was to start a song then click "MENU" and "HOLD". The Mini entered lockscreen before having the time to load everything, thus glitching the lockscreen until the next song.

The Mini is a slow device, maybe it's not required to do this workaround on lockscreen themes for the ipodclassic6g so that could explain why adwaitapod has choosen a ridiculous delay of 0.2 seconds.

The 3s delay seems to completely fix this issue at least on beginning of a song.

I can still reproduce this issue if I fast forward on a song then if I call instantly the lockscreen, it will also be glitched until next song. I don't know yet how to fix this, but it's very an uncommon way to glitch the lockscreen.